+++ renderer/RendererLib/RendererLib/src/ClientResourceUploadingManager.cpp
@@ -84,7 +84,7 @@ namespace ramses_internal
     void ClientResourceUploadingManager::uploadClientResources(const ResourceContentHashVector& resourcesToUpload)
     {
         UInt32 sizeUploaded = 0u;
-        for (UInt32 i = 0; i < resourcesToUpload.size(); ++i)
+        for (size_t i = 0; i < resourcesToUpload.size(); ++i)
         {
             const ResourceDescriptor& rd = m_clientResources.getResourceDescriptor(resourcesToUpload[i]);
             const UInt32 resourceSize = rd.resource.getResourceObject()->getDecompressedDataSize();
@@ -101,7 +101,7 @@ namespace ramses_internal
                 LOG_INFO_F(CONTEXT_RENDERER, [&](ramses_internal::StringOutputStream& logger)
                 {
                     logger << "Remaining resources in queue to upload:";
-                    for (UInt32 j = numUploaded; j < resourcesToUpload.size() && j < numUploaded + 10; ++j)
+                    for (size_t j = numUploaded; j < resourcesToUpload.size() && j < numUploaded + 10; ++j)
                     {
                         const ResourceDescriptor& interruptedRd = m_clientResources.getResourceDescriptor(resourcesToUpload[j]);
                         logger << " [" << interruptedRd.hash << "; " << EnumToString(interruptedRd.type) << "; " << interruptedRd.decompressedSize << " B]";
